requirements.txt
Appium-Python-Client==2.10.1
pytest==7.3.1
selenium~=4.9.1
behave==1.2.6